split running jobs portion of workspace detail into a new page
