+++ vite.config.js
@@ -13,10 +13,10 @@ const config = {
     rollupOptions: {
       input: {
         index: "./assets/src/scripts/index.js",
+        job_request_create: "./assets/src/scripts/job_request_create.js",
         main: "./assets/src/scripts/main.js",
         project_create: "./assets/src/scripts/project_create.js",
         workspace_create: "./assets/src/scripts/workspace_create.js",
-        workspace_detail: "./assets/src/scripts/workspace_detail.js",
         "outputs-viewer": "./assets/src/scripts/outputs-viewer/index.jsx",
       },
     },
